[YS-531] fix: 모바일 내가 쓴 글 클릭 시 바텀시트가 닫혀 모달이 보이지 않는 현상 해결
